+++ resources/views/studentsview/challenges.blade.php
@@ -8,7 +8,7 @@
        @endif
        <div class="p-2">
        @foreach(array_reverse($challenges) as $challenge)
-              <show-challenge :full="true" :prev-rating="{{ $challenge['rated'] }}" :challenge="{{ json_encode($challenge) }}" :edit="false" :admin="false" code="{{ $class->code }}"></show-challenge>
+              <show-challenge class="mb-3" :full="true" :prev-rating="{{ $challenge['rated'] }}" :challenge="{{ json_encode($challenge) }}" :edit="false" :admin="false" code="{{ $class->code }}"></show-challenge>
        @endforeach
        </div>
 @endsection